WE HAVE LIFTOFF!!!!!!
I just saw an rc prompt.

There were two lingering problems.

In Ken C, an int is 64 bits and a long is 32. In gcc,
quite the opposite. So that need a fix in Mach. Since assembly
depends on the size, I explicitly state the size as uint64_t.

Second, we changed the Label struct and it's now 16 quads, not two.
But I forget to fix the offset in the syscall entry. That's fixed.

And we have a prompt.

Nice work everyone. Now to get rid of all the debug prints. Have a
great day.

+++ sys/src/9/k10/dat.h
@@ -257,8 +257,8 @@ enum {
 struct Mach
 {
 	/* WARNING! Known to assembly! */
-	int	machno;			/* physical id of processor */
-	uintptr_t	splpc;			/* pc of last caller to splhi */
+	uint64_t	machno;			/* physical id of processor */
+	uint64_t	splpc;			/* pc of last caller to splhi */
 
 	Proc*	proc;			/* current process on this processor */
 	uintptr_t	stack;
